Added a new command "/list_admin" to list  my content, games, and books
