+++ bot.py
@@ -2,7 +2,7 @@ from config import bot
 from commands.add_movie import add_movie
 from commands.add_game import add_game
 from commands.add_book import add_book
-from commands.list_items import list_items
+from commands.list_items import list_user_items, list_admin_items
 from commands.add_many_movies import add_many_movies
 from commands.add_many_games import add_many_games
 from commands.start import send_welcome
@@ -14,7 +14,8 @@ from commands.admin import admin_panel
 bot.message_handler(commands=['add_movie'])(add_movie)
 bot.message_handler(commands=['add_game'])(add_game)
 bot.message_handler(commands=['add_book'])(add_book)
-bot.message_handler(commands=['list'])(list_items)
+bot.message_handler(commands=['list'])(list_user_items)
+bot.message_handler(commands=['list_admin'])(list_admin_items)
 bot.message_handler(commands=['add_many_movies'])(add_many_movies)
 bot.message_handler(commands=['add_many_games'])(add_many_games)
 bot.message_handler(commands=['start'])(send_welcome)
