+++ commands/start.py
@@ -22,6 +22,7 @@ def send_welcome(message):
         
         Команди для перегляду:
         /list - список контенту,ігор, книг
+        /list_admin - список контенту,ігор, книг адміна
 
         Команди для адміністрування:
         /admin - адмін панель
