+++ todo.md
@@ -8,7 +8,7 @@
 - [-] **Editing command**: `6` 🔵🟡🔴
 - [-] **Deleting command**: `6` 🔵🟡🔴
 - [x] **Input Database Limit**: `8` ⚪
-- [ ] **Command for output my data**: `7` ⚪
+- [x] **Command for output my data**: `7` ⚪
 - [ ] **Txt Export**: `1` ⚪
 - [x] **Restrict commands adding many media**: `8` 🔵🟡
 - [ ] **Admin Panel**: `5` ⚪
