+++ variables/globals.py
@@ -1,0 +1,1 @@
+query_dict = {}
